Don't show configuration error when Tor HS is not available
When we configure a Tor hidden service and haven't yet run Tor at least
once, then Tor hidden service hostname and private key are not
available.  This situation requires running Tor atleast once and is not
a configuration error.  When Tor runs, it see the hidden service
configuration and creates the Hidden service information.