Don't load Debian backup files as module configuration
When modules were removed, Debian packaging also had to remove them to
avoid errors.  When Debian guidelines are following and correspoding deb
helper is used, it creates .dpkg-* files in the module configuration
directory.  Plinth tries to load these which causes problems.  We should
consider loading files with a known extension in future.