owncloud: Do Apache reload instead of restart
Restarting Apache breaks existing connections.  This includes the connection
on which the Plinth's user has made the ownCloud enable/disable request.  This
leads to an almost certain 'Connection Interupted' message to the client when
changes to ownCloud are submitted.

On the other hand, 'reload' is sufficient and 'restart' is not required.  It is
also faster and recommended by 'a2enconf' and 'a2disconf' commands.