owncloud: Perform setup synchronously
Earlier, installation of the ownCloud package was performed as part of
setup process.  It took time and hence the action was done
asynchronously.  Now, however, package manager takes care of this.  To
avoid any potential race conditions with the user accessing ownCloud
before it is setup, perform the setup synchronously.