add Dockerfile for development
This adds a Dockerfile so you don't have to install all of Plinth's
required debian packages on your host machine in order to run it.

Let me know if you have any suggestions.. I just thought this could
be a good alternative to VirtualBox / KVM. It also provides a place
to document what's actually needed to get Plinth running.

I've added some instructions on how to run this to the INSTALL file.

    docker build .
    docker run -p 127.0.0.1:443:443 [image-hash]