Use icon for applications button intead of text
The concern that people are unable to find a way to get back to
applications is a correct one.  The recent changes does fix the problem.

However, it takes a step back in the design.  The 'Applications' text is
too prominent.  I propose that we replace it with the commonly used
icon.  Featured in mobile interfaces and on Google home page, users
should have no difficulty identifing it as applications icon.

After experimenting with the placement of the icon, it seems more
asthetically pleasing to have it at the place of 'Applications' text and
not at the center etc.