packages: Fix issue with handling some errors
I got an error when one of the sources in my sources.list is untrusted.
The error code was available as 'gpg-error' and more description was
available.  Our untested code to handle that threw an exception in that
case.  This patch fixes it.  I have tested with the error I got as the
error was easily reproducible.