config: Update dev configuration default url path
I find that I am almost always during development runs using ./run
--debug --no-daemon --server_dir=/plinth .  The daemon part is gone due
to recent changes.  But without specifying --server_dir it will run on /
and since it is not the same as production run, it change this to
/plinth.  I am assuming this also the case with other people.  So,
change the default to /plinth.  I believe this reduces the entry to
development barrier by a tiny bit.