Issue 496, status message improvement:
If no update is done because no update needs to be done, the success status needs to be written.
Because maybe the DNS record is up to date when the action script is executed for the first time.
In this case no success message was written and it looks like DynDNS is not executed.