letsencrypt: Depend on the new certbot package
Instead of the transitional package 'letsencrypt'.

I believe, there is no need to rename the entire module to certbo since
letsencrypt is stil a more apt name for the module.  Let's Encrypt is a
strong brand that people know.  The service is still called Let's
Encrypt.  Only the automation tool has been renamed to certbot to avoid
confusion of the service name.