sso: Increase timeout to 60 minutes
This is a temporary measure since no permanent solution was found for the cases
where auth_pubtkt cookie refresh fails when an application sends POST requests
only during the time a cookie refresh must happen. e.g. transmission and tt-rss

This at least makes the above applications usable for an hour. And then the user
must refresh the application on their web browser. This does not affect the
mobile apps.

The only other option for now is to disable SSO for the above applications till
we implement a better SSO solution which is undesirable since our users are
already used to the SSO feature.

Signed-off-by: Joseph Nuthalapati <njoseph@thoughtworks.com>
Reviewed-by: James Valleroy <jvalleroy@mailbox.org>