vagrant: Vagrantfile changes for ease of development
Moved some frequently required operations to the Vagrantfile.

Newcomers to the project almost always face the issue of port 8000 not being
available. Disabling system Plinth and running only the development version
solves this issue.

Automatic upgrades are a frequent blocker for development and an unnecessary
annoyance on a development setup.

- Update documentation about vagrant file changes

- Remove plinth restart step

Signed-off-by: Joseph Nuthalapati <njoseph@thoughtworks.com>
Reviewed-by: James Valleroy <jvalleroy@mailbox.org>