backups: Fix issue with showing exports from disks without labels
When a disk does not have a label, its device path is used. This contains a '/'
which by default is not encoded by 'urlencode' filter in Django templates. This
leads to the expected URL not matching the regex in urls.py. Fix this by
ensuring that '/' is also encoded by 'urlencode'.

Reviewed-by: James Valleroy <jvalleroy@mailbox.org>