searx: Set safe_search to Moderate by default
I remember the default being Moderate earlier. Maybe the Debian package
or upstream changed it to None. To be safe, we should set it in our
installation script too.

One problem I observed with default setting being None is that any
queries performed with this setting will be cached and a stricter
setting will not apply on the same search query. For example, if you
searched for "computer" with the default setting of None, there will be
some NSFW images returned by SearX. Setting safe_search to Moderate or
Strict later will have no effect on the search query "computer".

Signed-off-by: Joseph Nuthalapati <njoseph@riseup.net>
Reviewed-by: James Valleroy <jvalleroy@mailbox.org>